requirements.txt
Django<=3.0
psycopg2<=2.8
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1 
djangorestframework_simplejwt==5.0.0
django-rest-passwordreset   
django-allauth==0.48.0
django-ckeditor==6.2.0
celery==5.0.5
redis==3.5.3
django-cors-headers==3.11.0
Pillow==9.0.1
bs4==0.0.1
fcm-django<1
django-push-notifications==3.0.0
django-solo==2.0.0
firebase_admin==2.8.0   
coreapi
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
oauthlib
requests
requests-oauthlib
django-rest-auth==0.9.5